Separate history into its own package
This will make it easier to import this functionality into
gpbackup_manager
