+++ Makefile
@@ -37,7 +37,7 @@ lint :
 		gometalinter --config=gometalinter.config -s vendor ./...
 
 unit :
-		ginkgo -r -keepGoing -randomizeSuites -noisySkippings=false -randomizeAllSpecs backup restore utils testutils 2>&1
+		ginkgo -r -keepGoing -randomizeSuites -noisySkippings=false -randomizeAllSpecs backup restore utils backup_history testutils 2>&1
 
 integration :
 		ginkgo -r -randomizeSuites -noisySkippings=false -randomizeAllSpecs integration 2>&1
