+++ backup_history/backup_history_suite_test.go
@@ -1,0 +1,13 @@
+package backup_history_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestBackupHistory(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "BackupHistory Suite")
+}
