+++ backup_history/history.go
@@ -1,4 +1,6 @@
-package utils
+package backup_history
+
+//TODO: change package name to conform to Go standards
 
 import (
 	"sort"
@@ -7,16 +9,17 @@ import (
 	"github.com/greenplum-db/gp-common-go-libs/gplog"
 	"github.com/greenplum-db/gp-common-go-libs/iohelper"
 	"github.com/greenplum-db/gp-common-go-libs/operating"
+	"github.com/greenplum-db/gpbackup/utils"
 	"github.com/nightlyone/lockfile"
 	"gopkg.in/yaml.v2"
 )
 
 type History struct {
-	BackupConfigs []BackupConfig
+	BackupConfigs []utils.BackupConfig
 }
 
 func NewHistory(filename string) *History {
-	history := &History{BackupConfigs: make([]BackupConfig, 0)}
+	history := &History{BackupConfigs: make([]utils.BackupConfig, 0)}
 	if historyFileExists := iohelper.FileExistsAndIsReadable(filename); historyFileExists {
 		contents, err := operating.System.ReadFile(filename)
 
@@ -27,14 +30,14 @@ func NewHistory(filename string) *History {
 	return history
 }
 
-func (history *History) AddBackupConfig(backupConfig *BackupConfig) {
+func (history *History) AddBackupConfig(backupConfig *utils.BackupConfig) {
 	history.BackupConfigs = append(history.BackupConfigs, *backupConfig)
 	sort.Slice(history.BackupConfigs, func(i, j int) bool {
 		return history.BackupConfigs[i].Timestamp > history.BackupConfigs[j].Timestamp
 	})
 }
 
-func WriteBackupHistory(historyFilePath string, currentBackupConfig *BackupConfig) {
+func WriteBackupHistory(historyFilePath string, currentBackupConfig *utils.BackupConfig) {
 	lock := lockHistoryFile()
 	defer func() {
 		_ = lock.Unlock()
@@ -69,7 +72,7 @@ func (history *History) writeToFileAndMakeReadOnly(filename string) {
 	historyFile := iohelper.MustOpenFileForWriting(filename)
 	historyFileContents, err := yaml.Marshal(history)
 	gplog.FatalOnError(err)
-	MustPrintBytes(historyFile, historyFileContents)
+	utils.MustPrintBytes(historyFile, historyFileContents)
 	err = operating.System.Chmod(filename, 0444)
 	gplog.FatalOnError(err)
 }
