+++ gometalinter.config
@@ -6,6 +6,7 @@
     "comment on exported",
     "should not use dot imports",
     "don't use ALL_CAPS in Go names; use CamelCase",
-	"and that stutters"
+	"and that stutters",
+	"don't use an underscore in package name"
   ]
 }
